requirements.txt
colorama
colorist
datetime
requests
tls-client
websocket-client
uuid
typing_extensions
customtkinter
pyinstaller
pillow
